LayoutTests/http/tests/inspector/dom/site-isolated-node.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<!doctype html>
<html>
<head>
<script type="text/javascript" src="../resources/inspector-test.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("CommandLineAPI.$0.cross-frame");

    let mainFrame = WI.networkManager.mainFrame;
    let childFrames = Array.from(mainFrame.childFrameCollection);
    InspectorTest.expectEqual(childFrames.length, 1, "Page should have a subframe.");

    let iframeNodeId;

    suite.addTestCase({
        name: "AttemptCrossFrame$0AccessFromMainFrame",
        description: "Should not be able to access $0 node in different domain subframe from the main frame.",
        test(resolve, reject) {
            InspectorTest.log("Setting $0 to node within subframe.");
            DOMAgent.setInspectedNode(iframeNodeId);
            RuntimeAgent.evaluate.invoke({expression: "$0", includeCommandLineAPI: true}, (error, _, wasThrown) => {    
                InspectorTest.assert(!error, "Should not be a protocol error.");
                InspectorTest.assert(!wasThrown, "Should not be an exception.");
                resolve();
            });
        }
    });

    WI.domManager.requestDocument((documentNode) => {
        documentNode.querySelector("iframe#myframe", (nodeId) => {
            iframeNodeId = nodeId;
            suite.runTestCasesAndFinish();
        });
    });
}
</script>
</head>
<body>
<iframe id="myframe" src="http://localhost:8000/inspector/page/resources/test-page.html" onload="runTest()"></iframe>
</body>
</html>
